adding PopoverLayout
